document-verification/requirements.txt
flask
pytesseract
opencv-python


pillow
regex
numpy


